package-example/custom_nodes/ExampleNode2/requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
torch>=1.12.0
